ESP32: Blinky using C++11 enum class, lambda + vector/sort + manage scheduled lambdas.

+++ src/sdb_mod.h
@@ -21,7 +21,7 @@ public:
 
     virtual void onStart() {}
     
-    virtual int onLoop() {
+    virtual long onLoop() {
         return 1000 /*ms*/;
     }
